KV eval: adaptive early exit + configurable HF attention impl (eager=0.72)
evaluate_humaneval_evalplus.py:
- --kv_attn_impl {eager,sdpa,flash_attention_2}: force the HF model's attention
  kernel. eager gives the highest pass@1 (0.720) of the kernel-noise band
  (flash/sdpa 0.701, fp32 0.695) vs the no-cache torchtitan reference 0.707 --
  the ~1-4 problem spread is fused-kernel numerical noise, not a real gap.
- KV path already does adaptive early exit + avg_loops (prior commit).

Full-164 no-system threshold=1 attention/precision sweep slurms (flash/eager/
sdpa/fp32/no-prefill) + the DCP->HF bridge validation (convert_to_hf).

+++ torchtitan/models/ouro/config_registry.py
@@ -63,7 +63,10 @@ def ouro_1_4b() -> Trainer.Config:
             folder="/projects/bdjz/jli37/checkpoints",
             interval=100,
             last_save_model_only=False,
-            export_dtype="float16",
+            # The released Ouro checkpoint is bf16. Keep HF exports in bf16 too
+            # so KV-cache evals do not compare fp16-exported weights against the
+            # TorchTitan/DCP path.
+            export_dtype="bfloat16",
         ),
         activation_checkpoint=ActivationCheckpointConfig(
             mode="selective",
